==> src-code/2. Data Preprocessing.ipynb <==
!pip install -q nlpretext loguru

import pandas as pd
import numpy as np
from nlpretext import Preprocessor
from nlpretext.basic.preprocess import (normalize_whitespace, remove_punct, 
                                        remove_eol_characters, remove_stopwords, 
                                        lower_text, unpack_english_contractions,remove_accents,
                                        remove_multiple_spaces_and_strip_text,replace_urls,
                                       replace_emails,replace_currency_symbols)

from nlpretext.social.preprocess import (remove_mentions, remove_hashtag, 
                                         remove_emoji,remove_html_tags)

from nlpretext.token.preprocess import (remove_tokens_with_nonletters)

import logging
import time
import datetime

import warnings
warnings.filterwarnings("ignore")

logging.basicConfig(
    format='%(asctime)s %(levelname)-8s %(message)s',
    datefmt='%Y-%m-%d %H:%M:%S',
    level=logging.INFO, 
    filename='./logs/Data Preprocessing.log')   

logging.info("==========================================================================================================")
logging.info("Data Preprocessing Started ")

train_df = pd.read_csv('./input/train.csv')
test_df=pd.read_csv('./input/test.csv')
validation_df=pd.read_csv('./input/validation.csv')

logging.info(f"Train Shape {train_df.shape}, Test Shape {test_df.shape}, Validation Shape {validation_df.shape}")

print('Article:\n', train_df.iloc[0]['article'][:2000])
print('Summary:\n', train_df.iloc[0]['highlights'])

logging.info(f"Setting up preprocessing pipeline")

preprocessor = Preprocessor()
preprocessor.pipe(unpack_english_contractions)
preprocessor.pipe(normalize_whitespace)
preprocessor.pipe(remove_mentions)
preprocessor.pipe(remove_hashtag)
preprocessor.pipe(remove_emoji)
preprocessor.pipe(remove_punct, args={'marks': '])\'\",;:-([?'})
preprocessor.pipe(remove_accents)
preprocessor.pipe(replace_currency_symbols)
preprocessor.pipe(replace_emails)
preprocessor.pipe(replace_urls)
preprocessor.pipe(remove_html_tags)
preprocessor.pipe(remove_mentions)
preprocessor.pipe(remove_hashtag)
preprocessor.pipe(remove_emoji)

sample_train_df = train_df
sample_train_df.head()

logging.info(f"Cleaning Train dataset")
sample_train_df['article_cleaned'] = sample_train_df['article'].apply(preprocessor.run)

print('Article:\n', sample_train_df.iloc[4]['article'])
print('Article Cleaned:\n', sample_train_df.iloc[4]['article_cleaned'])

sample_train_df = sample_train_df.drop(['article'], axis=1)
sample_train_df.rename(columns = {'article_cleaned':'article'}, inplace = True) 

#Save preprocessed train data for fine tuning models
sample_train_df.to_csv('./input/train_cleaned.csv')
logging.info(f"Cleaned Train dataset saved into './input/train_cleaned.csv' file ")

logging.info(f"Cleaning Test dataset")
sample_test_df = test_df
sample_test_df.head()

sample_test_df['article_cleaned'] = sample_test_df['article'].apply(preprocessor.run)

print('Article:\n', sample_test_df.iloc[4]['article'])
print('Article Cleaned:\n', sample_test_df.iloc[4]['article_cleaned'])

sample_test_df = sample_test_df.drop(['article'], axis=1)
sample_test_df.rename(columns = {'article_cleaned':'article'}, inplace = True) 

#Save preprocessed train data for fine tuning models
sample_test_df.to_csv('./input/test_cleaned.csv')
logging.info(f"Cleaned Test dataset saved into './input/test_cleaned.csv' file ")

logging.info(f"Cleaning Validation dataset")
sample_validation_df = validation_df
sample_validation_df.head()

sample_validation_df['article_cleaned'] = sample_validation_df['article'].apply(preprocessor.run)

print('Article:\n', sample_validation_df.iloc[4]['article'])
print('Article Cleaned:\n', sample_validation_df.iloc[4]['article_cleaned'])

sample_validation_df = sample_validation_df.drop(['article'], axis=1)
sample_validation_df.rename(columns = {'article_cleaned':'article'}, inplace = True) 

#Save preprocessed train data for fine tuning models
sample_validation_df.to_csv('./input/validation_cleaned.csv')
logging.info(f"Cleaned Validation dataset saved into './input/validation_cleaned.csv' file ")

logging.info("Data Preprocessing Completed ")

df = pd.read_csv('./input/train_cleaned.csv')
df.head(3)
