==> src-code/3.2 Finetune BERT CNN.ipynb <==
!pip install datasets
!pip install evaluate
!pip install rouge_score
!pip install nltk
!pip install -U accelerate
!pip install -U transformers

from google.colab import drive
drive.mount('/content/drive')

%%capture
import datasets
import transformers
import pandas as pd
import numpy as np
from datasets import load_dataset
from transformers import AutoTokenizer
import torch
from transformers import Seq2SeqTrainingArguments, Seq2SeqTrainer
from transformers import EncoderDecoderModel
import logging
import time
import datetime

import warnings
warnings.filterwarnings("ignore")

#device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
#torch.cuda.empty_cache()

### Fine Tuning Pre Trained Model

logging.basicConfig(
    format='%(asctime)s %(levelname)-8s %(message)s',
    datefmt='%Y-%m-%d %H:%M:%S',
    level=logging.INFO,
    filename='/content/drive/My Drive/logs/BERT_Fine_Tuning.log',force=True)

logging.info("==========================================================================================================")
logging.info("BERT Fine Tuning Start ")

def roundTS(startTime, endTime):
    return round((endTime -startTime),4)


logging.info("Loading CNN Dataset ...")
train_data = datasets.load_dataset("cnn_dailymail", "3.0.0", split="train[:25%]", ignore_verifications=True)
val_data = datasets.load_dataset("cnn_dailymail", "3.0.0", split="validation[:3%]", ignore_verifications=True)
test_data = datasets.load_dataset("cnn_dailymail", "3.0.0", split="test[:1%]", ignore_verifications=True)

from transformers import BertTokenizerFast

tokenizer = BertTokenizerFast.from_pretrained("bert-base-uncased")
tokenizer.bos_token = tokenizer.cls_token
tokenizer.eos_token = tokenizer.sep_token

batch_size=12  # change to 16 for full training
encoder_max_length=512
decoder_max_length=128

def process_data_to_model_inputs(batch):
  # tokenize the inputs and labels
  inputs = tokenizer(batch["article"], padding="max_length", truncation=True, max_length=encoder_max_length)
  outputs = tokenizer(batch["highlights"], padding="max_length", truncation=True, max_length=decoder_max_length)

  batch["input_ids"] = inputs.input_ids
  batch["attention_mask"] = inputs.attention_mask
  batch["decoder_input_ids"] = outputs.input_ids
  batch["decoder_attention_mask"] = outputs.attention_mask
  batch["labels"] = outputs.input_ids.copy()

  # because BERT automatically shifts the labels, the labels correspond exactly to `decoder_input_ids`.
  # We have to make sure that the PAD token is ignored
  batch["labels"] = [[-100 if token == tokenizer.pad_token_id else token for token in labels] for labels in batch["labels"]]

  return batch

train_data = train_data.map(
    process_data_to_model_inputs,
    batched=True,
    batch_size=batch_size,
    remove_columns=["article", "highlights", "id"]
)
train_data.set_format(
    type="torch", columns=["input_ids", "attention_mask", "decoder_input_ids", "decoder_attention_mask", "labels"],
)


val_data = val_data.map(
    process_data_to_model_inputs,
    batched=True,
    batch_size=batch_size,
    remove_columns=["article", "highlights", "id"]
)
val_data.set_format(
    type="torch", columns=["input_ids", "attention_mask", "decoder_input_ids", "decoder_attention_mask", "labels"],
)

bert2bert = EncoderDecoderModel.from_encoder_decoder_pretrained("bert-base-uncased", "bert-base-uncased")

# set special tokens
bert2bert.config.decoder_start_token_id = tokenizer.bos_token_id
bert2bert.config.eos_token_id = tokenizer.eos_token_id
bert2bert.config.pad_token_id = tokenizer.pad_token_id

# sensible parameters for beam search
bert2bert.config.vocab_size = bert2bert.config.decoder.vocab_size
bert2bert.config.max_length = 60
bert2bert.config.min_length = 10
bert2bert.config.no_repeat_ngram_size = 3
bert2bert.config.early_stopping = True
bert2bert.config.length_penalty = 2.0
bert2bert.config.num_beams = 4

import evaluate
rouge = evaluate.load("rouge")
def compute_metrics(eval_pred):
    predictions, labels = eval_pred
    decoded_preds = tokenizer.batch_decode(predictions, skip_special_tokens=True)
    labels = np.where(labels != -100, labels, tokenizer.pad_token_id)
    decoded_labels = tokenizer.batch_decode(labels, skip_special_tokens=True)

    result = rouge.compute(predictions=decoded_preds, references=decoded_labels, use_stemmer=True)

    prediction_lens = [np.count_nonzero(pred != tokenizer.pad_token_id) for pred in predictions]
    result["gen_len"] = np.mean(prediction_lens)

    return {k: round(v, 4) for k, v in result.items()}



#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
from huggingface_hub import notebook_login
notebook_login()

st_bert_ft=time.time()

# set training arguments - these params are not really tuned, feel free to change
training_args = Seq2SeqTrainingArguments(
    "bert-base-cnn-ft",
    #output_dir="/content/drive/My Drive/models/BERT",
    evaluation_strategy="steps",
    per_device_train_batch_size=batch_size,
    per_device_eval_batch_size=batch_size,
    predict_with_generate=True,
    logging_steps=100,  # set to 1000 for full training
    save_steps=1000,  # set to 500 for full training
    eval_steps=100,  # set to 8000 for full training
    warmup_steps=1,  # set to 2000 for full training
    max_steps=2000, # delete for full training
    overwrite_output_dir=True,
    save_total_limit=1,
    fp16=True,
    weight_decay=0.01,
    optim = "adamw_torch",
    learning_rate=1e-5,
    push_to_hub=True,
    load_best_model_at_end=True,
)

# instantiate trainer
trainer = Seq2SeqTrainer(
    model=bert2bert,
    tokenizer=tokenizer,
    args=training_args,
    compute_metrics=compute_metrics,
    train_dataset=train_data,
    eval_dataset=val_data,
)
trainer.train()
logging.info(f"BERT FT Model pushed to Huggingface Repo at location harish3742/bert-base-cnn-ft")


end_bert_ft=time.time()
logging.info(f"BERT FT Duration - {roundTS(st_bert_ft,end_bert_ft)} seconds")
print(f"BERT FT Duration - {roundTS(st_bert_ft,end_bert_ft)} seconds")

trainer_state = trainer.state.log_history
logging.info(f"BERT Model Training Stats:\n" )

model = trainer.model

def print_trainable_parameters(model):
    """
    Prints the number of trainable parameters in the model.
    """
    trainable_params = 0
    all_param = 0
    for _, param in model.named_parameters():
        all_param += param.numel()
        if param.requires_grad:
            trainable_params += param.numel()
    logging.info(
        f"trainable params: {trainable_params} \nall params: {all_param} \ntrainable %: {100 * trainable_params / all_param}"
    )
    print(
        f"trainable params: {trainable_params} \nall params: {all_param} \ntrainable %: {100 * trainable_params / all_param}"
    )

def count_trainable_params(model):
    return sum(p.numel() for p in model.parameters() if p.requires_grad)

print_trainable_parameters(model)
print(count_trainable_params(model))


import os
file_size = os.stat('/content/bert-base-cnn-ft/model.safetensors')
print("file size :", round((file_size.st_size/ (1024 * 1024)), 2), "mb")

!pip install einops

params_dict = {
    'Name': [],
    'Shape': [],
    'Parameters': [],
    'RequiresGrad': []
}

for name, param in model.named_parameters():
    params_dict['Name'].append(name)
    params_dict['Shape'].append(param.shape)
    params_dict['Parameters'].append(param.numel())
    params_dict['RequiresGrad'].append(param.requires_grad)

df = pd.DataFrame(params_dict)
df.to_csv("/content/drive/My Drive/models/BERT/BERT_FT_Params.csv")
#logging.info("Model Parameters Table", df.head(20))
df.head(10)

### Saving Model to  HuggingFace Repo

#trainer.push_to_hub("harish3742/bert-base-cnn-ft-1")

#tokenizer = BertTokenizer.from_pretrained("bert-base-uncased")
from transformers import BertTokenizer
tokenizer = BertTokenizer.from_pretrained("/content/bert-base-cnn-ft")
model = EncoderDecoderModel.from_pretrained("/content/bert-base-cnn-ft")
model.to("cuda")

logging.info("BERT Fine Tuning End ")
