==> src-code/3.6 Prompt Tune GPT CNN.ipynb <==
import os
from openai import OpenAI
import pandas as pd
import numpy as np
from datasets import load_metric
import evaluate
import matplotlib.pyplot as plt
import matplotlib.gridspec as gridspec
rouge_metric = load_metric('rouge')

AZURE_OPENAI_KEY = "[REPLACE KEY ]"
AZURE_OPENAI_ENDPOINT = "https://tmap-openai.openai.azure.com"

import os
from openai import AzureOpenAI

client = AzureOpenAI(
  azure_endpoint = "https://tmap-openai.openai.azure.com", 
  api_key="[REPLACE KEY ]",  
  api_version="2023-09-01-preview",
)

#print(client.models.list())

### GPT - ATS - Prompt Tuning Experiements

sample_test_df = pd.read_csv('./input/test.csv', nrows=100)
sample_test_df = sample_test_df.sample(20)

def generate_GPT_ATS_Summary(record, model, temp, max_tokens):
    summary = ""
    
    abs_pmt = f"""
                Your task is to create a concise, factual summary, 
                
                by selecting and combining key sentences from  the original text. 
                
                Text is delimited by triple backticks. 

                TEXT: ```{record['article']}```
            """
     try:
        response = client.chat.completions.create(
        model=model, 
        messages=[
            {"role": "system", "content": "You are a LLM trained by OpenAI."},
            {"role": "user", "content": abs_pmt}

        ],
        max_tokens = max_tokens,
        temperature = temp,
        n=1,        
        )
        summary = response.choices[0].message.content
    except Exception:
        print(Exception)
        summary = "ERROR"
    
    return summary


def generateAbsBaselineComplexPrompt(text):
    summary = ""
    
    abs_pmt = f"""
                As a LLM trained by OpenAI, create a concise and factualsummary, 
                
                by selecting and combining key sentences from the original text,
                
                while adhering to these guidelines:
                
                1.Craft a summary that is detailed, in-depth, while maintaining clarity and conciseness.
                
                2.Incorporate main ideas, incorporate essential information and focus on critical aspects.
                
                3.Rely strictly on the provided text, without including external information.
                
                Text is delimited by triple backticks TEXT: ```{text}```
                """
   
    
    #create a concise summary by selecting and combining key sentences from the original text : {text}"
    try:
        response = client.chat.completions.create(
        model="gpt4", # model = "deployment_name".
        messages=[
            {"role": "system", "content": "You are a LLM trained by OpenAI."},
            {"role": "user", "content": abs_pmt}

        ])
        summary = response.choices[0].message.content
    except Exception:
        print(Exception)
        summary = "ERROR"
    
    return summary





sample_test_df['gpt4_ats_low_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 0, 30), axis=1)
sample_test_df['gpt4_ats_med_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 1, 30), axis=1)
sample_test_df['gpt4_ats_high_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 2, 30), axis=1)

sample_test_df['gpt35_ats_low_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 0, 30), axis=1)
sample_test_df['gpt35_ats_med_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 1, 30), axis=1)
sample_test_df['gpt35_ats_high_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 2, 30), axis=1)

sample_test_df['gpt4_ats_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 0, 60), axis=1)
sample_test_df['gpt4_ats_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 1, 60), axis=1)
sample_test_df['gpt4_ats_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 2, 60), axis=1)

sample_test_df['gpt35_ats_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 0, 60), axis=1)
sample_test_df['gpt35_ats_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 1, 60), axis=1)
sample_test_df['gpt35_ats_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 2, 60), axis=1)

sample_test_df['gpt4_ats_low_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 0, 90), axis=1)
sample_test_df['gpt4_ats_med_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 1, 90), axis=1)
sample_test_df['gpt4_ats_high_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 2, 90), axis=1)

sample_test_df['gpt35_ats_low_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 0, 90), axis=1)
sample_test_df['gpt35_ats_med_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 1, 90), axis=1)
sample_test_df['gpt35_ats_high_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 2, 90), axis=1)

sample_test_df['gpt4_ats_low_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 0, 120), axis=1)
sample_test_df['gpt4_ats_med_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 1, 120), axis=1)
sample_test_df['gpt4_ats_high_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt4', 2, 120), axis=1)

sample_test_df['gpt35_ats_low_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 0, 120), axis=1)
sample_test_df['gpt35_ats_med_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 1, 120), axis=1)
sample_test_df['gpt35_ats_high_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_ATS_Summary(rec, 'gpt35', 2, 120), axis=1)



rouge_names = ["rouge1", "rouge2", "rougeL", "rougeLsum"]

models = [
          
          "gpt4_ats_low_temp_30_words",  "gpt4_ats_med_temp_30_words",  "gpt4_ats_high_temp_30_words",
          "gpt35_ats_low_temp_30_words",  "gpt35_ats_med_temp_30_words",  "gpt35_ats_high_temp_30_words",
          
          "gpt4_ats_low_temp_60_words",  "gpt4_ats_med_temp_60_words",  "gpt4_ats_high_temp_60_words",
          "gpt35_ats_low_temp_60_words",  "gpt35_ats_med_temp_60_words",  "gpt35_ats_high_temp_60_words",
          
          "gpt4_ats_low_temp_90_words",  "gpt4_ats_med_temp_90_words",  "gpt4_ats_high_temp_90_words",
          "gpt35_ats_low_temp_90_words",  "gpt35_ats_med_temp_90_words",  "gpt35_ats_high_temp_90_words",
          
          "gpt4_ats_low_temp_120_words",  "gpt4_ats_med_temp_120_words",  "gpt4_ats_high_temp_120_words",
          "gpt35_ats_low_temp_120_words",  "gpt35_ats_med_temp_120_words",  "gpt35_ats_high_temp_120_words",
         ]


records = []

for model_name in models:
    rouge_metric.add(prediction = sample_test_df[model_name], reference = sample_test_df['highlights'])
    score = rouge_metric.compute()
    rouge_dict = dict((rn, score[rn].mid.fmeasure ) for rn in rouge_names )
    print('rouge_dict ', rouge_dict )
    records.append(rouge_dict)

metrics_df = pd.DataFrame.from_records(records, index = models )
metrics_df


metrics_df.to_csv('./output/GPT_ATS_PT_Metrics.csv')

### GPT - ATS - Few Shots Inferencing Experiements

sample_test_df = pd.read_csv('./input/test.csv', nrows=100)

# Computation of Length of article and hughlights
def article_len(row):
    return len(row['article'].split())

def highlights_len(row):
    return len(row['highlights'].split())


sample_test_df['article_len'] = sample_test_df.apply(lambda r: article_len(r), axis= 1)
sample_test_df['highlights_len'] = sample_test_df.apply(lambda r: highlights_len(r), axis= 1)


sorted_by_article_size_df = sample_test_df.sort_values('article_len')

sample_test_df = sample_test_df[:20]

def generate_GPT_Abs_1_Shot(record, model, temp, max_tokens):
    summary = ""
    abs_pmt = f"""
                Your task is to create a concise, factual summary, 
                
                by selecting and combining key sentences from  the original text. 
                
                Text is delimited by triple backticks. 

                TEXT: ```{record['article']}```
            """
    
    #create a concise summary by selecting and combining key sentences from the original text : {text}"
    try:
        response = client.chat.completions.create(
        model=model, # model = "deployment_name".
        messages=[
            {"role": "system", "content": "You are a LLM trained by OpenAI."},
            {"role":"user","content":sorted_by_article_size_df.iloc[0]['article']},
            {"role":"assistant","content":sorted_by_article_size_df.iloc[0]['highlights']},
            {"role": "user", "content": abs_pmt},
            ],
        max_tokens = max_tokens,
        temperature = temp, #top_p=.9
        n=1,
        )
        summary = response.choices[0].message.content
    except Exception:
        print(Exception)
        summary = "ERROR"
    
    return summary

def generate_GPT_Abs_3_Shots(record, model, temp, max_tokens):
    summary = ""
    
    abs_pmt = f"""
                Your task is to create a concise, factual summary, 
                
                by selecting and combining key sentences from  the original text. 
                
                Text is delimited by triple backticks. 

                TEXT: ```{record['article']}```
            """
    try:
        response = client.chat.completions.create(
        model=model, 
        messages=[
            {"role": "system", "content": "You are a LLM trained by OpenAI."},

            {"role":"user","content":sorted_by_article_size_df.iloc[0]['article']},
            {"role":"assistant","content":sorted_by_article_size_df.iloc[0]['highlights']},
            
            {"role":"user","content":sorted_by_article_size_df.iloc[1]['article']},
            {"role":"assistant","content":sorted_by_article_size_df.iloc[1]['highlights']},

            {"role":"user","content":sorted_by_article_size_df.iloc[2]['article']},
            {"role":"assistant","content":sorted_by_article_size_df.iloc[2]['highlights']},
            
            {"role": "user", "content": abs_pmt},
            
        ],
        max_tokens = max_tokens,
        temperature = temp,
        n=1,
        )
        summary = response.choices[0].message.content
    except Exception as e:
        print(e)
        summary = "ERROR"
    
    return summary

def generate_GPT_Abs_5_Shots(record, model, temp, max_tokens):
    summary = ""
       
    abs_pmt = f"""
                Your task is to create a concise, factual summary, 

                by selecting and combining key sentences from  the original text. 

                Text is delimited by triple backticks. 

                TEXT: ```{record['article']}```
            """
    #print('ARTICLE ', record['article'][:200])
        #create a concise summary by selecting and combining key sentences from the original text : {text}"
    try:
        response = client.chat.completions.create(
            model= model, # model = "deployment_name".
            messages=[
                {"role": "system", "content": "You are a LLM trained by OpenAI."},

                {"role":"user","content":sorted_by_article_size_df.iloc[0]['article']},
                {"role":"assistant","content":sorted_by_article_size_df.iloc[0]['highlights']},

                {"role":"user","content":sorted_by_article_size_df.iloc[1]['article']},
                {"role":"assistant","content":sorted_by_article_size_df.iloc[1]['highlights']},

                {"role":"user","content":sorted_by_article_size_df.iloc[2]['article']},
                {"role":"assistant","content":sorted_by_article_size_df.iloc[2]['highlights']},

                {"role":"user","content":sorted_by_article_size_df.iloc[3]['article']},
                {"role":"assistant","content":sorted_by_article_size_df.iloc[3]['highlights']},

                {"role":"user","content":sorted_by_article_size_df.iloc[4]['article']},
                {"role":"assistant","content":sorted_by_article_size_df.iloc[4]['highlights']},

                {"role": "user", "content": abs_pmt},

            ],
            max_tokens = max_tokens,
            temperature = temp, #top_p=.9
            n=1,
            )
        #print("Summary - ", response.choices[0].message.content)
        summary = response.choices[0].message.content
    except Exception as e:
        print(e)
        summary = "ERROR"
  
    return summary

#generate_GPT_Abs_5_Shots(sample_test_df.iloc[0], 'gpt4', 2, 60)


# Apply the function to each row of the dataframe
sample_test_df['gpt4-abs-5_shots_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_5_Shots(rec, 'gpt4', 0, 60), axis=1)
sample_test_df['gpt4-abs-5_shots_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_5_Shots(rec, 'gpt4', 1, 60), axis=1)
sample_test_df['gpt4-abs-5_shots_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_5_Shots(rec, 'gpt4', 2, 60), axis=1)

sample_test_df['gpt35-abs-5_shots_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_5_Shots(rec, 'gpt35', 0, 60), axis=1)
sample_test_df['gpt35-abs-5_shots_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_5_Shots(rec, 'gpt35', 1, 60), axis=1)
sample_test_df['gpt35-abs-5_shots_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_5_Shots(rec, 'gpt35', 2, 60), axis=1)

sample_test_df['gpt4-abs-3_shots_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_3_Shots(rec, 'gpt4', 0, 60), axis=1)
sample_test_df['gpt4-abs-3_shots_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_3_Shots(rec, 'gpt4', 1, 60), axis=1)
sample_test_df['gpt4-abs-3_shots_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_3_Shots(rec, 'gpt4', 2, 60), axis=1)

sample_test_df['gpt35-abs-3_shots_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_3_Shots(rec, 'gpt35', 0, 60), axis=1)
sample_test_df['gpt35-abs-3_shots_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_3_Shots(rec, 'gpt35', 1, 60), axis=1)
sample_test_df['gpt35-abs-3_shots_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_3_Shots(rec, 'gpt35', 2, 60), axis=1)

sample_test_df['gpt4-abs-1_shot_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_1_Shot(rec, 'gpt4', 0, 60), axis=1)
sample_test_df['gpt4-abs-1_shot_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_1_Shot(rec, 'gpt4', 1, 60), axis=1)
sample_test_df['gpt4-abs-1_shot_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_1_Shot(rec, 'gpt4', 2, 60), axis=1)

sample_test_df['gpt35-abs-1_shot_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_1_Shot(rec, 'gpt35', 0, 60), axis=1)
sample_test_df['gpt35-abs-1_shot_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_1_Shot(rec, 'gpt35', 1, 60), axis=1)
sample_test_df['gpt35-abs-1_shot_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Abs_1_Shot(rec, 'gpt35', 2, 60), axis=1)


rouge_names = ["rouge1", "rouge2", "rougeL", "rougeLsum"]
models = [

          'gpt4-abs-5_shots_low_temp_60_words', 
          'gpt4-abs-5_shots_med_temp_60_words' , 
          'gpt4-abs-5_shots_high_temp_60_words',
          'gpt35-abs-5_shots_low_temp_60_words', 
          'gpt35-abs-5_shots_med_temp_60_words', 
          'gpt35-abs-5_shots_high_temp_60_words',

          'gpt4-abs-3_shots_low_temp_60_words', 
          'gpt4-abs-3_shots_med_temp_60_words' , 
          'gpt4-abs-3_shots_high_temp_60_words',
          'gpt35-abs-3_shots_low_temp_60_words', 
          'gpt35-abs-3_shots_med_temp_60_words', 
          'gpt35-abs-3_shots_high_temp_60_words',
    
          'gpt4-abs-1_shot_low_temp_60_words', 
          'gpt4-abs-1_shot_med_temp_60_words' , 
          'gpt4-abs-1_shot_high_temp_60_words',
          'gpt35-abs-1_shot_low_temp_60_words', 
          'gpt35-abs-1_shot_med_temp_60_words', 
          'gpt35-abs-1_shot_high_temp_60_words',
         ]

records = []

for model_name in models:
    rouge_metric.add(prediction = sample_test_df[model_name], reference = sample_test_df['highlights'])
    score = rouge_metric.compute()
    rouge_dict = dict((rn, score[rn].mid.fmeasure ) for rn in rouge_names )
    print('rouge_dict ', rouge_dict )
    records.append(rouge_dict)

metrics_df = pd.DataFrame.from_records(records, index = models )
metrics_df

metrics_df.to_csv("./output/GPT_ATS_FS_Metrics.csv")

### GPT ETS Prompt Tuning

sample_test_df = pd.read_csv('./input/test.csv', nrows=100)
sample_test_df = sample_test_df[:20]

def generate_GPT_Ext_Summary(record, model, temp, max_tokens):
    summary = ""
    
    ext_pmt = f"""
                Your task is to Generate extractive summary from the original text, 
                
                by identifying 4-5 key sentences from the original text and 
                
                DO NOT rewrite or paraphrase the content.
                
                Text is delimited by triple backticks. 

                TEXT: ```{record['article']}```
            """
    try:
        response = client.chat.completions.create(
        model=model, 
        messages=[
            {"role": "system", "content": "You are a LLM trained by OpenAI."},
            {"role": "user", "content": ext_pmt}

        ],
        max_tokens = max_tokens,
        temperature = temp,
        n=1,
        )
        summary = response.choices[0].message.content
    except Exception as e:
        print(e)
        summary = "ERROR"
    
    return summary

# Apply the function to each row of the dataframe

sample_test_df['gpt4_ets_low_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 0, 30), axis=1)
sample_test_df['gpt4_ets_med_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 1, 30), axis=1)
sample_test_df['gpt4-ets_high_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 2, 30), axis=1)

sample_test_df['gpt35_ets_low_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 0, 30), axis=1)
sample_test_df['gpt35_ets_med_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 1, 30), axis=1)
sample_test_df['gpt35-ets_high_temp_30_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 2, 30), axis=1)

sample_test_df['gpt4_ets_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 0, 60), axis=1)
sample_test_df['gpt4_ets_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 1, 60), axis=1)
sample_test_df['gpt4-ets_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 2, 60), axis=1)

sample_test_df['gpt35_ets_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 0, 60), axis=1)
sample_test_df['gpt35_ets_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 1, 60), axis=1)
sample_test_df['gpt35-ets_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 2, 60), axis=1)

sample_test_df['gpt4_ets_low_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 0, 90), axis=1)
sample_test_df['gpt4_ets_med_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 1, 90), axis=1)
sample_test_df['gpt4-ets_high_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 2, 90), axis=1)

sample_test_df['gpt35_ets_low_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 0, 90), axis=1)
sample_test_df['gpt35_ets_med_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 1, 90), axis=1)
sample_test_df['gpt35-ets_high_temp_90_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 2, 90), axis=1)

sample_test_df['gpt4_ets_low_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 0, 120), axis=1)
sample_test_df['gpt4_ets_med_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 1, 120), axis=1)
sample_test_df['gpt4-ets_high_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt4', 2, 120), axis=1)

sample_test_df['gpt35_ets_low_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 0, 120), axis=1)
sample_test_df['gpt35_ets_med_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 1, 120), axis=1)
sample_test_df['gpt35-ets_high_temp_120_words'] = sample_test_df.apply(lambda rec: generate_GPT_Ext_Summary(rec, 'gpt35', 2, 120), axis=1)


rouge_names = ["rouge1", "rouge2", "rougeL", "rougeLsum"]
models = [

          'gpt4_ets_low_temp_30_words', 
          'gpt4_ets_med_temp_30_words' , 
          'gpt4-ets_high_temp_30_words',
          'gpt35_ets_low_temp_30_words', 
          'gpt35_ets_med_temp_30_words', 
          'gpt35-ets_high_temp_30_words',
    
          'gpt4_ets_low_temp_60_words', 
          'gpt4_ets_med_temp_60_words' , 
          'gpt4-ets_high_temp_60_words',
          'gpt35_ets_low_temp_60_words', 
          'gpt35_ets_med_temp_60_words', 
          'gpt35-ets_high_temp_60_words',

          'gpt4_ets_low_temp_90_words', 
          'gpt4_ets_med_temp_90_words' , 
          'gpt4-ets_high_temp_90_words',
          'gpt35_ets_low_temp_90_words', 
          'gpt35_ets_med_temp_90_words', 
          'gpt35-ets_high_temp_90_words',

          'gpt4_ets_low_temp_120_words', 
          'gpt4_ets_med_temp_120_words' , 
          'gpt4-ets_high_temp_120_words',
          'gpt35_ets_low_temp_120_words', 
          'gpt35_ets_med_temp_120_words', 
          'gpt35-ets_high_temp_120_words',
         ]

records = []

for model_name in models:
    rouge_metric.add(prediction = sample_test_df[model_name], reference = sample_test_df['highlights'])
    score = rouge_metric.compute()
    rouge_dict = dict((rn, score[rn].mid.fmeasure ) for rn in rouge_names )
    print('rouge_dict ', rouge_dict )
    records.append(rouge_dict)

metrics_df = pd.DataFrame.from_records(records, index = models )
metrics_df

metrics_df.to_csv('./output/GPT_ETS_PT_Metrics.csv')

### GPT ETS Few Shots Tuning

sample_test_df = pd.read_csv('./input/test.csv', nrows=100)

# Computation of Length of article and hughlights
def article_len(row):
    return len(row['article'].split())

def highlights_len(row):
    return len(row['highlights'].split())


sample_test_df['article_len'] = sample_test_df.apply(lambda r: article_len(r), axis= 1)
sample_test_df['highlights_len'] = sample_test_df.apply(lambda r: highlights_len(r), axis= 1)

sorted_by_article_size_df = sample_test_df.sort_values('article_len')

sample_test_df = sample_test_df[:20]

def generate_GPT_ETS_1_Shot(record, model, temp, max_tokens):
    summary = ""
    ext_pmt = f"""
                Your task is to Generate extractive summary from the original text, 
                
                by identifying 4-5 key sentences from the original text and 
                
                DO NOT rewrite or paraphrase the content.
                
                Text is delimited by triple backticks. 

                TEXT: ```{record['article']}```
            """
    
    #create a concise summary by selecting and combining key sentences from the original text : {text}"
    try:
        response = client.chat.completions.create(
        model=model, # model = "deployment_name".
        messages=[
            {"role": "system", "content": "You are a LLM trained by OpenAI."},
            {"role":"user","content":sorted_by_article_size_df.iloc[0]['article']},
            {"role":"assistant","content":sorted_by_article_size_df.iloc[0]['highlights']},
            {"role": "user", "content": ext_pmt},
            ],
        max_tokens = max_tokens,
        temperature = temp, #top_p=.9
        n=1,
        )
        summary = response.choices[0].message.content
    except Exception:
        print(Exception)
        summary = "ERROR"
    
    return summary

def generate_GPT_ETS_3_Shots(record, model, temp, max_tokens):
    summary = ""
    
    ext_pmt = f"""
                Your task is to Generate extractive summary from the original text, 
                
                by identifying 4-5 key sentences from the original text and 
                
                DO NOT rewrite or paraphrase the content.
                
                Text is delimited by triple backticks. 

                TEXT: ```{record['article']}```
            """
    
    #create a concise summary by selecting and combining key sentences from the original text : {text}"
    try:
        response = client.chat.completions.create(
        model=model, # model = "deployment_name".
        messages=[
            {"role": "system", "content": "You are a LLM trained by OpenAI."},

            {"role":"user","content":sorted_by_article_size_df.iloc[0]['article']},
            {"role":"assistant","content":sorted_by_article_size_df.iloc[0]['highlights']},
            
            {"role":"user","content":sorted_by_article_size_df.iloc[1]['article']},
            {"role":"assistant","content":sorted_by_article_size_df.iloc[1]['highlights']},

            {"role":"user","content":sorted_by_article_size_df.iloc[2]['article']},
            {"role":"assistant","content":sorted_by_article_size_df.iloc[2]['highlights']},
            
            {"role": "user", "content": ext_pmt},
            
        ],
        max_tokens = max_tokens,
        temperature = temp, #top_p=.9
        n=1,
        )
        summary = response.choices[0].message.content
    except Exception as e:
        print(e)
        summary = "ERROR"
    
    return summary

def generate_GPT_ETS_5_Shots(record, model, temp, max_tokens):
    summary = ""
       
    ext_pmt = f"""
                Your task is to Generate extractive summary from the original text, 
                
                by identifying 4-5 key sentences from the original text and 
                
                DO NOT rewrite or paraphrase the content.
                
                Text is delimited by triple backticks. 

                TEXT: ```{record['article']}```
            """
    #print('ARTICLE ', record['article'][:200])
        #create a concise summary by selecting and combining key sentences from the original text : {text}"
    try:
        response = client.chat.completions.create(
            model= model, # model = "deployment_name".
            messages=[
                {"role": "system", "content": "You are a LLM trained by OpenAI."},

                {"role":"user","content":sorted_by_article_size_df.iloc[0]['article']},
                {"role":"assistant","content":sorted_by_article_size_df.iloc[0]['highlights']},

                {"role":"user","content":sorted_by_article_size_df.iloc[1]['article']},
                {"role":"assistant","content":sorted_by_article_size_df.iloc[1]['highlights']},

                {"role":"user","content":sorted_by_article_size_df.iloc[2]['article']},
                {"role":"assistant","content":sorted_by_article_size_df.iloc[2]['highlights']},

                {"role":"user","content":sorted_by_article_size_df.iloc[3]['article']},
                {"role":"assistant","content":sorted_by_article_size_df.iloc[3]['highlights']},

                {"role":"user","content":sorted_by_article_size_df.iloc[4]['article']},
                {"role":"assistant","content":sorted_by_article_size_df.iloc[4]['highlights']},

                {"role": "user", "content": ext_pmt},

            ],
            max_tokens = max_tokens,
            temperature = temp, 
            top_p= 0.9,
            presence_penalty=0.01,
            n=1,
            )
        #print("Summary - ", response.choices[0].message.content)
        summary = response.choices[0].message.content
    except Exception as e:
        print(e)
        summary = "ERROR"
  
    return summary

# Apply the function to each row of the dataframe
sample_test_df['gpt4-ets-5_shots_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_5_Shots(rec, 'gpt4', 0, 60), axis=1)
sample_test_df['gpt4-ets-5_shots_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_5_Shots(rec, 'gpt4', 1, 60), axis=1)
sample_test_df['gpt4-ets-5_shots_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_5_Shots(rec, 'gpt4', 2, 60), axis=1)

sample_test_df['gpt35-ets-5_shots_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_5_Shots(rec, 'gpt35', 0, 60), axis=1)
sample_test_df['gpt35-ets-5_shots_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_5_Shots(rec, 'gpt35', 1, 60), axis=1)
sample_test_df['gpt35-ets-5_shots_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_5_Shots(rec, 'gpt35', 2, 60), axis=1)

sample_test_df['gpt4-ets-3_shots_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_3_Shots(rec, 'gpt4', 0, 60), axis=1)
sample_test_df['gpt4-ets-3_shots_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_3_Shots(rec, 'gpt4', 1, 60), axis=1)
sample_test_df['gpt4-ets-3_shots_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_3_Shots(rec, 'gpt4', 2, 60), axis=1)

sample_test_df['gpt35-ets-3_shots_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_3_Shots(rec, 'gpt35', 0, 60), axis=1)
sample_test_df['gpt35-ets-3_shots_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_3_Shots(rec, 'gpt35', 1, 60), axis=1)
sample_test_df['gpt35-ets-3_shots_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_3_Shots(rec, 'gpt35', 2, 60), axis=1)

sample_test_df['gpt4-ets-1_shot_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_1_Shot(rec, 'gpt4', 0, 60), axis=1)
sample_test_df['gpt4-ets-1_shot_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_1_Shot(rec, 'gpt4', 1, 60), axis=1)
sample_test_df['gpt4-ets-1_shot_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_1_Shot(rec, 'gpt4', 2, 60), axis=1)

sample_test_df['gpt35-ets-1_shot_low_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_1_Shot(rec, 'gpt35', 0, 60), axis=1)
sample_test_df['gpt35-ets-1_shot_med_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_1_Shot(rec, 'gpt35', 1, 60), axis=1)
sample_test_df['gpt35-ets-1_shot_high_temp_60_words'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_1_Shot(rec, 'gpt35', 2, 60), axis=1)


rouge_names = ["rouge1", "rouge2", "rougeL", "rougeLsum"]
models = [

          'gpt4-ets-5_shots_low_temp_60_words', 
          'gpt4-ets-5_shots_med_temp_60_words' , 
          'gpt4-ets-5_shots_high_temp_60_words',
          'gpt35-ets-5_shots_low_temp_60_words', 
          'gpt35-ets-5_shots_med_temp_60_words' , 
          'gpt35-ets-5_shots_high_temp_60_words',

          'gpt4-ets-3_shots_low_temp_60_words', 
          'gpt4-ets-3_shots_med_temp_60_words' , 
          'gpt4-ets-3_shots_high_temp_60_words',
          'gpt35-ets-3_shots_low_temp_60_words', 
          'gpt35-ets-3_shots_med_temp_60_words' , 
          'gpt35-ets-3_shots_high_temp_60_words',
    
          'gpt4-ets-1_shot_low_temp_60_words', 
          'gpt4-ets-1_shot_med_temp_60_words' , 
          'gpt4-ets-1_shot_high_temp_60_words',
          'gpt35-ets-1_shot_low_temp_60_words', 
          'gpt35-ets-1_shot_med_temp_60_words' , 
          'gpt35-ets-1_shot_high_temp_60_words',
         ]

records = []

for model_name in models:
    rouge_metric.add(prediction = sample_test_df[model_name], reference = sample_test_df['highlights'])
    score = rouge_metric.compute()
    rouge_dict = dict((rn, score[rn].mid.fmeasure ) for rn in rouge_names )
    print('rouge_dict ', rouge_dict )
    records.append(rouge_dict)

metrics_df = pd.DataFrame.from_records(records, index = models )
metrics_df

metrics_df

metrics_df.to_csv("./output/GPT_ETS_FS_Metrics.csv")
