==> src-code/3.7 Plot GPT Prompt Tuning Metrics.ipynb <==
import pandas as pd
import numpy as np
#from rouge import Rouge
from datasets import load_metric
import evaluate
import matplotlib.pyplot as plt
import matplotlib.gridspec as gridspec
rouge_metric = load_metric('rouge')

### Plot GPT ATS Zero Shot Metrics

df = pd.read_csv('./output/GPT_ATS_PT_Metrics.csv')
df.rename(columns = {'Unnamed: 0':'GPT ATS Zero Shot Methods', "rouge1":"ROUGE 1", "rouge2":"ROUGE 2","rougeL":"ROUGE L", "rougeLsum": 'ROUGE LSUM'}, inplace = True) 
df.head(20)

### Plot GPT ETS Zero Shot Metrics

df = pd.read_csv('./output/GPT_ETS_PT_Metrics.csv')
df.rename(columns = {'Unnamed: 0':'GPT ETS Zero Shot Methods', "rouge1":"ROUGE 1", "rouge2":"ROUGE 2","rougeL":"ROUGE L", "rougeLsum": 'ROUGE LSUM'}, inplace = True) 
df.head(20)

### Plot GPT ATS Few Shot Metrics

df = pd.read_csv('./output/GPT_ATS_FS_Metrics.csv')
df.rename(columns = {'Unnamed: 0':'GPT ATS Few Shot Methods', "rouge1":"ROUGE 1", "rouge2":"ROUGE 2","rougeL":"ROUGE L", "rougeLsum": 'ROUGE LSUM'}, inplace = True) 
  
df.head(20)

### Plot GPT ETS Few Shot Metrics

df = pd.read_csv('./output/GPT_ETS_FS_Metrics.csv')
df.rename(columns = {'Unnamed: 0':'GPT ETS Few Shot Methods', "rouge1":"ROUGE 1", "rouge2":"ROUGE 2","rougeL":"ROUGE L", "rougeLsum": 'ROUGE LSUM'}, inplace = True) 
df.head(20)
