==> src-code/4. Extractive Text Summary Generation.ipynb <==
import pandas as pd
import numpy as np
import nltk
from nltk.tokenize import sent_tokenize

nltk.download("punkt")

import os
from openai import OpenAI,AzureOpenAI

from summarizer import Summarizer

from datasets import load_dataset, load_from_disk, load_metric
from transformers import pipeline, set_seed,AutoModelForSeq2SeqLM, AutoTokenizer, BertTokenizer,EncoderDecoderModel
import logging
import time
import torch

import warnings
warnings.filterwarnings("ignore")

logging.basicConfig(
    format='%(asctime)s %(levelname)-8s %(message)s',
    datefmt='%Y-%m-%d %H:%M:%S',
    level=logging.INFO, 
    filename='./logs/Extractive Text Summary Generation.log')    

logging.info("==========================================================================================================")
logging.info("Extractive Text Summarization Start ")

NO_OF_TEST_RECORDS = 100
logging.info(f"No of test records - {NO_OF_TEST_RECORDS}")

start_time_ets=time.time()

# Load sample records from test data
sample_test_df = pd.read_csv('./input/test_cleaned.csv', nrows=NO_OF_TEST_RECORDS)
sample_test_df.head(2)

summaries = {}

duration={}

# Zero Shot Extractive Summarization

#### Baseline Model (Lead 5)

def generate_ext_baseline(text):
    return "\n".join(sent_tokenize(text)[:5]) # First 5 sentences

def roundTS(startTime, endTime):
    return round((endTime -startTime),4)

def avgTimePerRecord(startTime, endTime, no_of_recs):
     return round((endTime -startTime)/no_of_recs ,4)

logging.info("Generating Baseline Extractive Summaries...")
st_baseline_ets=time.time()
sample_test_df['baseline-ext'] = sample_test_df['article'].apply(generate_ext_baseline)
end_baseline_ets=time.time()
logging.info(f"Baseline ETS Duration - {roundTS(st_baseline_ets, end_baseline_ets)} seconds")    
sample_test_df['baseline-ext'][:2]

duration['Baseline'] = avgTimePerRecord(st_baseline_ets, end_baseline_ets, NO_OF_TEST_RECORDS)

print("Article:\n",sample_test_df.iloc[0]['article'][:512])
print("Summary:\n",sample_test_df.iloc[0]['baseline-ext'])

#### BERT Extractive Summarization

model = Summarizer()
def generate_ext_bert_zs(text):
    result = model(text, min_length=40)
    return "".join(result)

logging.info("Generating BERT Extractive Summaries (Before Fine Tuning) ...")    
st_bert_zs_ets=time.time()
sample_test_df['bert-base-ext-zs'] = sample_test_df['article'].apply(generate_ext_bert_zs)
end_bert_zs_ets=time.time()
logging.info(f"BERT ZS ETS Duration - {roundTS(st_bert_zs_ets,end_bert_zs_ets)} seconds")    
sample_test_df.head(3)

duration['BERT'] = avgTimePerRecord(st_bert_zs_ets, end_bert_zs_ets, NO_OF_TEST_RECORDS)

print("Article\n",sample_test_df.iloc[0]['article'])
print("Summary\n",sample_test_df.iloc[0]['bert-base-ext-zs'])

#### T5 Extractive Summarization

def generate_ext_t5_small_zs(text):
    pipe = pipeline('summarization', model = 't5-small', min_length=40)
    pipe_out = pipe(text)
    return 'n'.join(sent_tokenize(pipe_out[0]['summary_text']))

logging.info("Generating T5 Extractive Summaries ...")    
st_t5_zs_ets=time.time()
sample_test_df['t5-small-ext-zs'] = sample_test_df['article'].apply(generate_ext_t5_small_zs)
end_t5_zs_ets=time.time()
logging.info(f"T5 ZS ETS Duration - {roundTS(st_t5_zs_ets, end_t5_zs_ets) } seconds")    
sample_test_df.head()


duration['T5'] = avgTimePerRecord(st_t5_zs_ets, end_t5_zs_ets, NO_OF_TEST_RECORDS)

print("Article\n", sample_test_df.iloc[0]['article'])
print("Summary\n", sample_test_df.iloc[0]['t5-small-ext-zs'])

#### GPT 3.5 Extractive Summarization

client = OpenAI(
    api_key='[API KEY]',
)
AZURE_OPENAI_KEY = "[API KEY]"
AZURE_OPENAI_ENDPOINT = "https://tmap-openai.openai.azure.com"

client = AzureOpenAI(
  azure_endpoint = "https://tmap-openai.openai.azure.com/", 
  api_key="[API KEY]",  
  api_version="2023-09-15-preview")

def generate_ext_gpt_35_zs(text):
    GPT35_ETS_PMT = "Generate extractive summary from the original text : " + text
    summary = ""
    try:
        response = client.chat.completions.create(
            model="gpt35", 
            messages=[
                {"role": "system", "content": "You are a LLM trained by OpenAI."},
                {"role": "user", "content": GPT35_ETS_PMT}
            ]
        )
        summary = response.choices[0].message.content
    except Exception as e:
        logging.info(f"Exception occured - {e}")
        summary = "ERROR"
    return summary


logging.info("Generating GPT3.5 Extractive Summaries (Before Prompt Tuning) ...") 
st_gpt35_zs_ets=time.time()
sample_test_df['gpt-3.5-ext-zs'] = sample_test_df['article'].apply(generate_ext_gpt_35_zs)
end_gpt35_zs_ets=time.time()
logging.info(f"GPT35 ZS ETS Duration - {roundTS(st_gpt35_zs_ets, end_gpt35_zs_ets)} seconds")    
sample_test_df.head(3)

duration['GPT35'] = avgTimePerRecord(st_gpt35_zs_ets, end_gpt35_zs_ets, NO_OF_TEST_RECORDS)

print("Article\n",sample_test_df.iloc[0]['article'])
print("Summary\n",sample_test_df.iloc[0]['gpt-3.5-ext-zs'])

#### GPT 4 Extractive Summarization

client = AzureOpenAI(
  azure_endpoint = "https://tmap-openai.openai.azure.com/", 
  api_key="[API KEY]",  
  api_version="2023-09-15-preview")

def generate_ext_gpt_4_zs(text):
    GPT40_ETS_PMT = "create a concise summary by selecting and combining key sentences from the original text : " + text
    summary = ""
    try:
        response = client.chat.completions.create(
            model="gpt4", 
            messages=[
                {"role": "system", "content": "You are a LLM trained by OpenAI."},
                {"role": "user", "content": GPT40_ETS_PMT}

            ]
        )
        summary = response.choices[0].message.content
    except Exception as e:
        logging.info(f"Exception Occured - {e}")
        summary = "ERROR"
        
    return summary

logging.info("Generating GPT4 Extractive Summaries (Before Prompt Tuning) ...") 
st_gpt4_zs_ets=time.time()
sample_test_df['gpt-4-ext-zs'] = sample_test_df['article'].apply(generate_ext_gpt_4_zs)
end_gpt4_zs_ets=time.time()
logging.info(f"GPT4 ZS ETS Duration - {roundTS(st_gpt4_zs_ets, end_gpt4_zs_ets)} seconds")    
sample_test_df.head(3)

end_gpt4_zs_ets, st_gpt4_zs_ets

duration['GPT4'] = avgTimePerRecord(st_gpt4_zs_ets, end_gpt4_zs_ets, NO_OF_TEST_RECORDS)

print("Article\n",sample_test_df.iloc[0]['article'])
print("Summary\n",sample_test_df.iloc[0]['gpt-4-ext-zs'])

# Fine Tuned - Extractractive Text Summarization

###### Prerequisite - Fine Tuned BERT, T5 Pretrained Models 

#### BERT Extractive Summarization

sample_test_df

device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')

bert_tokenizer = BertTokenizer.from_pretrained("./models/bert-base-cnn-finetuned")
bert_model_ft = EncoderDecoderModel.from_pretrained("./models/bert-base-cnn-finetuned")
bert_model_ft.to(device)


def generate_ext_bert_ft(text):
    # cut off at BERT max length 512
    inputs = bert_tokenizer([text], padding="max_length", truncation=True, max_length=512, return_tensors="pt")
    input_ids = inputs.input_ids.to(device)
    attention_mask = inputs.attention_mask.to(device)
    output = bert_model_ft.generate(input_ids, attention_mask=attention_mask)
    return bert_tokenizer.decode(output[0], skip_special_tokens=True)

logging.info("Generating BERT Extractive Summaries (After Fine Tuning) ...")
st_bert_ft_ets=time.time()
sample_test_df['bert-base-ext-ft'] = sample_test_df['article'].apply(generate_ext_bert_ft)
end_bert_ft_ets=time.time()
logging.info(f"BERT FT ETS Duration - {roundTS(st_bert_ft_ets, end_bert_ft_ets)} seconds")    
sample_test_df.head(3)

duration['BERT-FT'] = avgTimePerRecord(st_bert_ft_ets, end_bert_ft_ets,NO_OF_TEST_RECORDS)

#### T5 Extractive Summarization

print("T5 Doesnot support Extractive Summarization !")

#### GPT 3.5 Extractive Summarization

# Computation of Length of article and hughlights
def article_len(row):
    return len(row['article'].split())

def highlights_len(row):
    return len(row['highlights'].split())


sample_test_df['article_len'] = sample_test_df.apply(lambda r: article_len(r), axis= 1)
sample_test_df['highlights_len'] = sample_test_df.apply(lambda r: highlights_len(r), axis= 1)

sorted_by_article_size_df = sample_test_df.sort_values('article_len')

def generate_GPT_ETS_1_Shot(record, model, temp, max_tokens):
    summary = ""
    ext_pmt = f"""
                Your task is to Generate extractive summary from the original text, 
                
                by identifying 4-5 key sentences from the original text and 
                
                DO NOT rewrite or paraphrase the content.
                
                Text is delimited by triple backticks. 

                TEXT: ```{record['article']}```
            """
    
    #create a concise summary by selecting and combining key sentences from the original text : {text}"
    try:
        response = client.chat.completions.create(
        model=model, # model = "deployment_name".
        messages=[
            {"role": "system", "content": "You are a LLM trained by OpenAI."},
            {"role":"user","content":sorted_by_article_size_df.iloc[0]['article']},
            {"role":"assistant","content":sorted_by_article_size_df.iloc[0]['highlights']},
            {"role": "user", "content": ext_pmt},
            ],
        max_tokens = max_tokens,
        temperature = temp, 
        n=1,
        )
        summary = response.choices[0].message.content
    except Exception as e:
        logging.info(f"Exception Occured - {e}")
        summary = "ERROR"
    
    return summary

logging.info("Generating GPT35 Extractive Summaries (After Prompt Tuning) ...") 
st_gpt35_ft_ets=time.time()
sample_test_df['gpt35-ext-ft'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_1_Shot(rec, 'gpt35', 0, 60), axis=1)
end_gpt35_ft_ets=time.time()
logging.info(f"GPT35 FT ETS Duration - {roundTS(st_gpt35_ft_ets, end_gpt35_ft_ets)} seconds")    
sample_test_df.head(3)

duration['GPT35-FT'] = avgTimePerRecord(st_gpt35_ft_ets, end_gpt35_ft_ets,NO_OF_TEST_RECORDS)

#### GPT 4 Extractive Summarization

logging.info("Generating GPT4 Extractive Summaries (After Prompt Tuning) ...") 
st_gpt4_ft_ets=time.time()
sample_test_df['gpt40-ext-ft'] = sample_test_df.apply(lambda rec: generate_GPT_ETS_1_Shot(rec, 'gpt4', 0, 60), axis=1)
end_gpt4_ft_ets=time.time()
logging.info(f"GPT4 FT ETS Duration - {roundTS(st_gpt4_ft_ets, end_gpt4_ft_ets)} seconds")   


duration['GPT4-FT'] = avgTimePerRecord(st_gpt4_ft_ets, end_gpt4_ft_ets,NO_OF_TEST_RECORDS)

sample_test_df=sample_test_df.drop(['article_len', 'highlights_len'], axis=1)

#### Saving the generated summaries for evaluation

end_time_ets=time.time()
logging.info(f"Total ETS Duration - {roundTS(start_time_ets, end_time_ets)} seconds")

duration

duration_df = pd.DataFrame(duration.items(), columns=['models', 'avg_inf_time'])
duration_df

logging.info(f"MAP - Average Inference Time Per Request - {duration}")

file_path = './output/ext-ts-final.csv'
logging.info(f"Summaries will be saved in this file : {file_path}")
sample_test_df.to_csv(file_path,  mode="w+")

duration_file_path = './output/ext-ts-duration-final.csv'
logging.info(f"Duration will be saved in this file : {duration_file_path}")
duration_df.to_csv(duration_file_path,  mode="w+")

sample_test_df.head(3)

logging.info("Extractive Text Summarization End ")
read_content = pd.read_csv(file_path)
read_content.head(3)
