==> src-code/6. Hybrid Text Summarization.ipynb <==
import pandas as pd
import numpy as np
import nltk
from nltk.tokenize import sent_tokenize
import time
nltk.download("punkt")

import os
from openai import OpenAI,AzureOpenAI

from datasets import load_dataset, load_from_disk, load_metric
from transformers import pipeline, set_seed,AutoModelForSeq2SeqLM, AutoTokenizer, BertTokenizer,EncoderDecoderModel

import logging
import time
import datetime

import torch

import warnings
warnings.filterwarnings("ignore")

logging.basicConfig(
    format='%(asctime)s %(levelname)-8s %(message)s',
    datefmt='%Y-%m-%d %H:%M:%S',
    level=logging.INFO, 
    filename='./logs/Hybrid Text Summary Generation.log')   

logging.info("==========================================================================================================")
logging.info("Hybrid Text Summarization Started ")

NO_OF_TEST_RECORDS = 100
logging.info(f"No of Test Records considered for analysis - {NO_OF_TEST_RECORDS}")

start_time_hts=time.time()

sample_test_df = pd.read_csv('./input/test_cleaned.csv', nrows=NO_OF_TEST_RECORDS)
sample_test_df.head()

summaries = {}

duration={}

# Baseline Hybrid Summarization

### Lead 3 + Seq2Seq (Ext - Abs Hybrid summarization)

logging.info("Lead 3 + Seq2Seq hybrid summarization")
def generate_ext_baseline(text):
    return "\n".join(sent_tokenize(text)[:3]) # First 3 sentences

pipe = pipeline("summarization", model="bert-base-uncased", max_new_tokens=80)

def generate_abs_seq2seq(text):
   pipe_out = pipe(text)
   return "\n".join(sent_tokenize(pipe_out[0]["summary_text"]))

def generate_hyb_baseline(text):
    temp_summary = generate_ext_baseline(text)
    summary = generate_abs_seq2seq(temp_summary)
    return summary

def roundTS(startTime, endTime):
    return round((endTime -startTime),4)

def avgTimePerRecord(startTime, endTime, no_of_recs):
     return round((endTime -startTime)/no_of_recs ,4)

logging.info("Generating Baseline Hybrid Summaries...")
st_baseline_hts=time.time()
sample_test_df['baseline-hyb'] = sample_test_df['article'].apply(generate_hyb_baseline)
end_baseline_hts=time.time()
logging.info(f"Baseline ATS Duration - {roundTS(st_baseline_hts, end_baseline_hts)} seconds") 

duration['Baseline'] = avgTimePerRecord(st_baseline_hts, end_baseline_hts, NO_OF_TEST_RECORDS)

sample_test_df.head()  

### Lead 3 + BERT (Ext - Abs Hybrid summarization)

logging.info("Lead 3 + BERT hybrid summarization")
device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
bert_tokenizer = BertTokenizer.from_pretrained("./models/bert-base-cnn-finetuned")
bert_model_ft = EncoderDecoderModel.from_pretrained("./models/bert-base-cnn-finetuned")
bert_model_ft.to(device)


def generate_abs_bert(text):
    # cut off at BERT max length 512
    inputs = bert_tokenizer([text], padding="max_length", truncation=True, max_length=512, return_tensors="pt")
    input_ids = inputs.input_ids.to(device)
    attention_mask = inputs.attention_mask.to(device)
    output = bert_model_ft.generate(input_ids, attention_mask=attention_mask)
    return bert_tokenizer.decode(output[0], skip_special_tokens=True)

def generate_hyb_bert(text):
    temp_summary = generate_ext_baseline(text)
    summary = generate_abs_bert(temp_summary)
    return summary

logging.info("Generating BERT Hybrid Summaries...")
st_bert_hyb_hts=time.time()
sample_test_df['bert-hyb'] = sample_test_df['article'].apply(generate_hyb_bert)
end_bert_hyb_hts=time.time()
logging.info(f"BERT HTS Duration - {roundTS(st_bert_hyb_hts, end_bert_hyb_hts)} seconds")  

sample_test_df.head() 

duration['bert-hyb'] = avgTimePerRecord(st_bert_hyb_hts, end_bert_hyb_hts, NO_OF_TEST_RECORDS)

### Lead 3 + T5 (Ext - Abs Hybrid summarization)

logging.info("Lead 3 + T5 hybrid summarization")
from transformers import pipeline
summarizer = pipeline("summarization", model="./models/t5-small-cnn-hf-finetuned")

def generate_abs_t5(text):
    response = summarizer(text)
    return response[0]['summary_text']


def generate_hyb_t5(text):
    temp_summary = generate_ext_baseline(text)
    summary = generate_abs_t5(temp_summary)
    return summary

logging.info("Generating T5 Hybrid Summaries...")
st_t5_hyb_hts=time.time()
sample_test_df['t5-hyb'] = sample_test_df['article'].apply(generate_hyb_t5)
end_t5_hyb_hts=time.time()
logging.info(f"T5 HTS Duration - {roundTS(st_t5_hyb_hts, end_t5_hyb_hts)} seconds")  


sample_test_df.head()  

duration['t5-hyb'] = avgTimePerRecord(st_t5_hyb_hts, end_t5_hyb_hts, NO_OF_TEST_RECORDS)

### Lead 3 + GPT 3.5 (Ext - Abs Hybrid summarization)

logging.info("Lead 3 + GPT3.5 hybrid summarization")
client = OpenAI(
    # This is` the default and can be omitted
    api_key='[API KEY]',
)
AZURE_OPENAI_KEY = "[API KEY]"
AZURE_OPENAI_ENDPOINT = "https://tmap-openai.openai.azure.com"

client = AzureOpenAI(
  azure_endpoint = "https://tmap-openai.openai.azure.com/", 
  api_key="[API KEY]",  
  api_version="2023-09-15-preview")

# Computation of Length of article and hughlights
def article_len(row):
    return len(row['article'].split())

def highlights_len(row):
    return len(row['highlights'].split())

sample_test_df['article_len'] = sample_test_df.apply(lambda r: article_len(r), axis= 1)
sample_test_df['highlights_len'] = sample_test_df.apply(lambda r: highlights_len(r), axis= 1)

sorted_by_article_size_df = sample_test_df.sort_values('article_len')

def generate_GPT_Hyb_1_Shot(temp_summary, model, temp, max_tokens):
    summary = ""
    abs_pmt = f"""
                Your task is to create a concise, factual summary, 
                
                by selecting and combining key sentences from  the original text. 
                
                Text is delimited by triple backticks. 

                TEXT: ```{temp_summary}```
            """
    
    #create a concise summary by selecting and combining key sentences from the original text : {text}"
    try:
        response = client.chat.completions.create(
        model=model, # model = "deployment_name".
        messages=[
            {"role": "system", "content": "You are a LLM trained by OpenAI."},
            {"role":"user","content":sorted_by_article_size_df.iloc[0]['article']},
            {"role":"assistant","content":sorted_by_article_size_df.iloc[0]['highlights']},
            {"role": "user", "content": abs_pmt},
            ],
        max_tokens = max_tokens,
        temperature = temp, #top_p=.9
        n=1,
        )
        summary = response.choices[0].message.content
    except Exception as e:
        logging.info("Exception occured - {e}") 
        summary = "ERROR"
    
    return summary

def generate_hyb_gpt35(record, model, temp, max_tokens):
    temp_summary = generate_ext_baseline(record['article'])
    summary = generate_GPT_Hyb_1_Shot(temp_summary,model, temp, max_tokens)
    return summary

logging.info("Generating GPT3.5 Hybrid Summaries...")
st_gpt35_hyb_hts=time.time()
sample_test_df['gpt35-hyb'] = sample_test_df.apply(lambda rec: generate_hyb_gpt35(rec, 'gpt35', 0, 60), axis=1)
end_gpt35_hyb_hts=time.time()
logging.info(f"GPT 3.5 HTS Duration - {roundTS(st_gpt35_hyb_hts, end_gpt35_hyb_hts)} seconds")  


duration['gpt35-hyb'] = avgTimePerRecord(st_gpt35_hyb_hts, end_gpt35_hyb_hts, NO_OF_TEST_RECORDS)

sample_test_df.head()  

### Lead 3 + GPT 4 (Ext - Abs Hybrid summarization)

logging.info("Lead 3 + GPT 4 hybrid summarization")
client = AzureOpenAI(
  azure_endpoint = "https://tmap-openai.openai.azure.com/", 
  api_key="[API KEY]",  
  api_version="2023-09-15-preview")

def generate_hyb_gpt4(record, model, temp, max_tokens):
    temp_summary = generate_ext_baseline(record['article'])
    summary = generate_GPT_Hyb_1_Shot(temp_summary,model, temp, max_tokens)
    return summary

#sample_test_df['gpt4-hyb'] = sample_test_df['article'].apply(generate_hyb_gpt4)
logging.info("Generating GPT 4 Hybrid Summaries...")
st_gpt4_hyb_hts=time.time()
sample_test_df['gpt4-hyb'] = sample_test_df.apply(lambda rec: generate_hyb_gpt35(rec, 'gpt4', 0, 60), axis=1)
end_gpt4_hyb_hts=time.time()
logging.info(f"GPT 4 HTS Duration - {roundTS(st_gpt4_hyb_hts, end_gpt4_hyb_hts)} seconds") 


duration['gpt4-hyb'] = avgTimePerRecord(st_gpt4_hyb_hts, end_gpt4_hyb_hts, NO_OF_TEST_RECORDS)

sample_test_df.head()  

end_time_hts=time.time()
logging.info(f"Total HTS Duration - {roundTS(start_time_hts, end_time_hts)} seconds")


duration

duration_df = pd.DataFrame(duration.items(), columns=['models', 'avg_inf_time'])
duration_df

logging.info(f"MAP - HTS Average Inference Time Per Request - {duration}")

duration_file_path = './output/hyb-ts-duration-final.csv'
logging.info(f"Duration will be saved in this file : {duration_file_path}")
duration_df.to_csv(duration_file_path,  mode="w+")

file_path = './output/hyb-ts-final.csv'
logging.info(f"Summaries will be saved in this file : {file_path}")

sample_test_df.head(5)

sample_test_df.to_csv(file_path,  mode="w+")

read_content = pd.read_csv(file_path)
read_content.tail(5)

logging.info("Hybrid Text Summarization Completed ")
